pico2w: address PR #92 review comments
- Delete APU drain_samples_into; local transport uses drain_samples_to closure
- Split render_menu_row and render_loading_row into focused sub-functions
- Eliminate FlashRomStager temporal dependency: filename stored at begin(),
  header committed automatically on final write_next_bank(), finish() removed
- Move state types into src/state/ with one module per state
- Rename cpu -> gameboy throughout RunningState and main()
