+++ platform/pico2w/src/state/running.rs
@@ -1,0 +1,63 @@
+use alloc::vec::Vec;
+
+use embassy_rp::peripherals::PIO0;
+use embassy_rp::pio_programs::i2s::PioI2sOut;
+use rustyboy_pico2w::audio::AudioBuffers;
+use rustyboy_pico2w::display::hw::GameDisplay;
+use rustyboy_pico2w::input::InputHandler;
+use rustyboy_pico2w::multicore::PicoGameBoy;
+
+use crate::{App, AppState, CYCLES_PER_FRAME, poll_once};
+use super::InGameMenuState;
+
+pub struct RunningState;
+
+impl RunningState {
+    pub async fn tick(
+        &mut self,
+        app: &mut App,
+        gameboy: &mut PicoGameBoy,
+        game_disp: &mut GameDisplay<'static>,
+        i2s: &mut PioI2sOut<'static, PIO0, 0>,
+        input: &mut InputHandler<'static>,
+        audio_buffers: &mut AudioBuffers,
+        audio_samples: &mut Vec<i16>,
+    ) {
+        // Scope the DMA futures so `game_disp` is free after the block.
+        let open_menu = {
+            let frame_buf = gameboy.published_scaled_frame();
+
+            let mut disp_future = core::pin::pin!(game_disp.send_frame_raw(frame_buf));
+            let _ = poll_once(disp_future.as_mut());
+
+            let (front_buf, back_buf) = audio_buffers.front_back_buffers();
+            let mut audio_future = core::pin::pin!(i2s.write(front_buf));
+            let _ = poll_once(audio_future.as_mut());
+
+            let frame_start = gameboy.cycle_counter();
+            while gameboy.cycle_counter().wrapping_sub(frame_start) < CYCLES_PER_FRAME {
+                gameboy.tick();
+            }
+
+            let (current_buttons, open_menu) = input.poll();
+            for (btn, pressed) in app.previous_buttons.diff(current_buttons) {
+                gameboy.set_button(btn, pressed);
+            }
+            app.previous_buttons = current_buttons;
+
+            gameboy.drain_audio_samples_into_i16(audio_samples);
+            audio_buffers.queue_next_frame_i16(audio_samples, back_buf);
+
+            disp_future.as_mut().await;
+            gameboy.release_scaled_frame();
+            audio_future.as_mut().await;
+
+            open_menu
+        };
+
+        if open_menu {
+            let next = InGameMenuState::new(game_disp, app).await;
+            app.transition_to(AppState::InGameMenu(next));
+        }
+    }
+}
